Add shared utils module and refactor scripts
