+++ email_video_runner.py
@@ -13,14 +13,20 @@ import logging
 import time
 import os
 import re
-import sqlite3
 import random
 from email.utils import parseaddr
 
 import yt_dlp
 
+from utils import (
+    DB_PATH,
+    initialize_database,
+    is_url_downloaded,
+    record_successful_download,
+    safe_filename,
+)
+
 # ---------------------- Constants ----------------------
-DB_PATH = "downloads.db"
 MAX_IMAP_RETRIES = int(os.environ.get("MAX_IMAP_RETRIES", "3"))
 IMAP_RETRY_DELAY = float(os.environ.get("IMAP_RETRY_DELAY", "1.0"))
 YTDLP_COOKIES_FILE = os.environ.get("YTDLP_COOKIES_FILE")
@@ -28,47 +34,10 @@ FACTCHECK_KEYWORD = "factcheck"
 FACTCHECK_RESPONSE_PREFIX = "[FACTCHECK - RESPONSE] - "  # New for loop prevention
 
 # ---------------------- DB Functions ----------------------
-def initialize_database(db_path: str = DB_PATH):
-    conn = sqlite3.connect(db_path)
-    cursor = conn.cursor()
-    cursor.execute("""
-        CREATE TABLE IF NOT EXISTS downloads (
-            id INTEGER PRIMARY KEY AUTOINCREMENT,
-            url TEXT UNIQUE NOT NULL,
-            title TEXT NOT NULL,
-            downloaded_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-        )
-    """)
-    conn.commit()
-    conn.close()
-
-
-def is_url_downloaded(url: str, db_path: str = DB_PATH) -> bool:
-    conn = sqlite3.connect(db_path)
-    cursor = conn.cursor()
-    cursor.execute("SELECT 1 FROM downloads WHERE url = ?", (url,))
-    result = cursor.fetchone()
-    conn.close()
-    return result is not None
-
-
-def record_successful_download(url: str, title: str, db_path: str = DB_PATH):
-    conn = sqlite3.connect(db_path)
-    cursor = conn.cursor()
-    try:
-        cursor.execute("INSERT OR IGNORE INTO downloads (url, title) VALUES (?, ?)", (url, title))
-        conn.commit()
-        logging.info(f"[i] Recorded download: {url} with title '{title}'")
-    except sqlite3.Error as e:
-        logging.error(f"[x] DB record error for {url}: {e}")
-    finally:
-        conn.close()
+# Provided by utils module
 
 # ---------------------- Filename Utility ----------------------
-def safe_filename(input_str: str, max_length: int = 150) -> str:
-    sanitized = re.sub(r'[\\/*?:"<>|]', '_', input_str.strip())
-    sanitized = sanitized.strip('. ').strip()[:max_length]
-    return sanitized or "untitled"
+# Provided by utils module
 
 # ---------------------- IMAP Reconnection ----------------------
 def reconnect_imap(imap_srv, email_addr: str, passwd: str, mailbox: str):
